ref(options): rename config-key helpers and Rust module off "runtime config"
Per review, these no longer deal with the legacy runtime config:

- Rust: rename the `sentry_options` module to `snuba_options` so it no
  longer shadows the external `sentry_options` crate; drop the leading-`::`
  disambiguation that shadowing required in lib.rs.
- Python: rename `_build_runtime_config_key` -> `_build_config_key` and
  `__deserialize_runtime_config_key` -> `__deserialize_config_key` (they
  build/parse the config's key regardless of backing store).
